Datasets

Phone data

LinRegOutliers.DataSets.phones — Constant

Phone data

Components

year::Integer: years from 1950 to 1973.

calls::Float64: phone calls (in millions).

Reference

P. J. Rousseeuw and A. M. Leroy (1987) Robust Regression & Outlier Detection. Wiley.

Hawkings & Bradu & Kass data

LinRegOutliers.DataSets.hbk — Constant

Hawkins & Bradu & Kass data

Components

x1::Float64: first independent variable.

x2::Float64: second independent variable.

x3::Float64: third independent variable.

y::Float64: dependent (response) variable.

Reference

Hawkins, D.M., Bradu, D., and Kass, G.V. (1984) Location of several outliers in multiple regression data using elemental sets. Technometrics 26, 197–208.

Animals data

Missing docstring.

Missing docstring for LinRegOutliers.animals. Check Documenter's build log for details.

Weight Loss data

LinRegOutliers.DataSets.weightloss — Constant

Weight loss data

Components

days::Integer: time in days since the start of the diet program.

weight::Float64: weight in kg.

Reference

 Venables, W. N. and Ripley, B. D. (1999) _Modern Applied
 Statistics with S-PLUS._ Third Edition. Springer.

Stack Loss data

LinRegOutliers.DataSets.stackloss — Constant

Stack loss data

Components

airflow::Float64: flow of cooling air (independent variable).

watertemp::Float64: cooling water inlet temperature (independent variable).

acidcond::Float64: concentration of acid (independent variable).

stackloss::Float64: stack loss (dependent variable).

Outliers

Observations 1, 3, 4, and 21 are outliers.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) _The New S Language_.  Wadsworth & Brooks/Cole.

Dodge, Y. (1996) The guinea pig of multiple regression. In: _Robust Statistics, Data Analysis, and Computer Intensive Methods;
In Honor of Peter Huber's 60th Birthday_, 1996, _Lecture Notes in Statistics_ *109*, Springer-Verlag, New York.

Hadi & Simonoff (1993) random data

LinRegOutliers.DataSets.hs93randomdata — Constant

Hadi & Simonoff (1993) Random data

Components

x1::Float64: Random values.

x2::Float64: Random values.

y::Float64: Random values (independent variable).

Outliers

Observations 1, 2, and 3 are outliers.

References

Hadi, Ali S., and Jeffrey S. Simonoff. "Procedures for the identification of multiple outliers in linear models." Journal of the American Statistical Association 88.424 (1993): 1264-1272.

Modified Wood Gravity data

LinRegOutliers.DataSets.woodgravity — Constant

Modified Wood Gravity Data

Components

x1::Float64: Random values.

x2::Float64: Random values.

x3::Float64: Random values.

x4::Float64: Random values.

x5::Float64: Random values.

y::Float64: Random values (independent variable).

References

P. J. Rousseeuw and A. M. Leroy (1987) Robust Regression and Outlier Detection. Wiley, p.243, table 8.

Scottish Hill Races data

LinRegOutliers.DataSets.hills — Constant

Scottish Hill Races Data

Components

dist::Array{Float64, 1}: Distance in miles (Independent).

climb::Array{Float64, 1}: Heights in feet (Independent).

time::Array{Float64, 1}: Record times in hours (Dependent).

Model

time ~ dist + climb

References

A.C. Atkinson (1986) Comment: Aspects of diagnostic regression analysis. Statistical Science 1, 397-402.

Soft Drink Delivery data

LinRegOutliers.DataSets.softdrinkdelivery — Constant

Soft Drink Delivery Data

Components

cases::Array{Float64, 1}: Independent variable.

distance::Array{Float64, 1}: Independent variable.

time::Array{Float64, 1}: Dependent variable.

Model

time ~ distance + cases

Reference

D. C. Montgomery and E. A. Peck (1992) Introduction to Regression Analysis. Wiley, New York.