LinRegOutliers.jl Documentation

LinRegOutliers.RegressionSetting — Type

struct RegressionSetting
    formula::FormulaTerm
    data::DataFrame
end

Immutable data structure for a regression setting.

Arguments

formula::FormulaTerm: A formula object describes the linear regression model.

data::DataFrame: DataFrame object holds the data

Notes

Implemented methods in this packages accepts linear models as RegressionSetting objects.
This objects holds the model formula and the data used in regression estimations.

Examples

julia> setting = RegressionSetting(@formula(calls ~ year), phones)
RegressionSetting(calls ~ year, 24×2 DataFrame
│ Row │ year  │ calls   │
│     │ Int64 │ Float64 │
├─────┼───────┼─────────┤
│ 1   │ 50    │ 4.4     │
│ 2   │ 51    │ 4.7     │
│ 3   │ 52    │ 4.7     │
│ 4   │ 53    │ 5.9     │
│ 5   │ 54    │ 6.6     │
│ 6   │ 55    │ 7.3     │
│ 7   │ 56    │ 8.1     │
│ 8   │ 57    │ 8.8     │
│ 9   │ 58    │ 10.6    │
│ 10  │ 59    │ 12.0    │
⋮
│ 14  │ 63    │ 21.2    │
│ 15  │ 64    │ 119.0   │
│ 16  │ 65    │ 124.0   │
│ 17  │ 66    │ 142.0   │
│ 18  │ 67    │ 159.0   │
│ 19  │ 68    │ 182.0   │
│ 20  │ 69    │ 212.0   │
│ 21  │ 70    │ 43.0    │
│ 22  │ 71    │ 24.0    │
│ 23  │ 72    │ 27.0    │
│ 24  │ 73    │ 29.0    │)

LinRegOutliers.adjustedResiduals — Method

adjustedResiduals(setting)

Calculate adjusted residuals for a given regression setting.

# Arguments:

setting::RegressionSetting: A regression setting object.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);
julia> adjustedResiduals(reg)
24-element Array{Float64,1}:
  13.486773572526268
   8.2307993473897
   2.774371467851612
  -1.3093999279776498
  -5.851901346871404
 -10.335509559699863
 -14.670907823058053
 -19.07911256736661
 -22.338710565623828
 -26.00786250934617
 -29.58187157605512
 -33.27523207616458
 -37.19977737822219
 -37.173743587631165
  57.781855070799956
  57.898085871534626
  71.47231139524963
  84.19185329435882
 103.37399662263209
 130.23557965295348
 -52.6720662600165
 -78.76891816539992
 -81.75736547266746
 -85.9914301855088

LinRegOutliers.applyColumns — Method

applyColumns(f, data)

Apply function f to each columns of data.

Arguments

f::Function: A function that takes a one dimensional array as argument.

data::Matrix: A Matrix object.

LinRegOutliers.applyColumns — Method

applyColumns(f, data)

Apply function f to each columns of data.

Arguments

f::Function: A function that takes a one dimensional array as argument.

data::DataFrame: A DataFrame object.

LinRegOutliers.asm2000 — Method

asm2000(setting)

Perform the Setan, Halim and Mohd (2000) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> asm2000(reg0001)
Dict{Any,Any} with 1 entry:
  "outliers" => [15, 16, 17, 18, 19, 20]

References

Setan, Halim, and Mohd Nor Mohamad. "Identifying multiple outliers in linear regression: Robust fit and clustering approach." (2000).

LinRegOutliers.atkinson94 — Method

    atkinson94(setting, iters, crit)

Runs the Atkinson94 algorithm to find out outliers using LMS method.

Arguments

setting::RegressionSetting: A regression setting object.

iters::Int: Number of random samples.

crit::Float64: Critical value for residuals

Examples

julia> reg = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> atkinson94(reg)
Dict{Any,Any} with 6 entries:
  "optimum_index"    => 10
  "residuals_matrix" => [0.0286208 0.0620609 … 0.0796249 0.0; 0.0397778 0.120547 … 0.118437 0.0397778; … ; 1.21133 1.80846 … 0.690327 4.14366; 1.61977 0.971592 … 0.616204 3.58098]
  "outliers"         => [1, 3, 4, 21]
  "objective"        => 0.799134
  "coef"             => [-38.3133, 0.745659, 0.432794, 0.0104587]
  "crit"             => 3.0

References

Atkinson, Anthony C. "Fast very robust methods for the detection of multiple outliers." Journal of the American Statistical Association 89.428 (1994): 1329-1339.

LinRegOutliers.atkinsonstalactiteplot — Method

Runs the atkinson94 algorithm and additionally plots the Stalactite text plot as described in the paper. Works for data with less than 100 points at the moment (since screen width of 80-120 is the standard).

References

Atkinson, Anthony C. "Fast very robust methods for the detection of multiple outliers." Journal of the American Statistical Association 89.428 (1994): 1329-1339.

LinRegOutliers.bacon — Method

    bacon(setting, m, method, alpha)

Run the BACON algorithm to detect outliers on regression data.

Arguments:

setting: RegressionSetting object with a formula and a dataset.

m: The number of elements to be included in the initial subset.

method: The distance method to use for selecting the points for initial subset

alpha: The quantile used for cutoff

Examples

julia> reg = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> bacon_multivariate_outlier_detection(reg, m=12)
4-element Array{Int64,1}:
  1
  3
  4
 21

References

Billor, Nedret, Ali S. Hadi, and Paul F. Velleman. "BACON: blocked adaptive computationally efficient outlier nominators." Computational statistics & data analysis 34.3 (2000): 279-298.

LinRegOutliers.bacon_multivariate_outlier_detection — Method

    bacon_multivariate_outlier_detection(X, m, method, alpha)

This function performs the outlier detection for multivariate data according to algorithm #3. This is used by algorithm #4 to compute the initial subset.

Arguments

X: The multivariate data matrix.

m: The minimum number of points to include in the initial subset

method: The distance method to use for selecting the points for initial subset

alpha: The quantile used for cutoff

LinRegOutliers.bacon_regression_initial_subset — Method

    bacon_regression_initial_subset(X, y, m, method, alpha)

This function computes the initial subset having at least m elements which are likely to be free of outliers used for the BACON algorithm.

Arguments:

X: The multivariate data matrix.

y: The response vector.

m: The minimum number of points to include in the initial subset

method: The distance method to use for selecting the points for initial subset

alpha: The quantile used for cutoff

LinRegOutliers.bch — Method

bch(setting; alpha = 0.05, maxiter = 1000, epsilon = 0.000001)

Perform the Billor & Chatterjee & Hadi (2006) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

alpha::Float64: Optional argument of the probability of rejecting the null hypothesis.

maxiter::Int: Maximum number of iterations for calculating iterative weighted least squares estimates.

epsilon::Float64: Accuracy for determining convergency.

Examples

julia> reg  = createRegressionSetting(@formula(calls ~ year), phones);
julia> Dict{Any,Any} with 7 entries:
"betas"                               => [-55.9205, 1.15572]
"squared.normalized.robust.distances" => [0.104671, 0.0865052, 0.0700692, 0.0553633, 0.0423875, 0.03…
"weights"                             => [0.00186158, 0.00952088, 0.0787321, 0.0787321, 0.0787321, 0…
"outliers"                            => [1, 14, 15, 16, 17, 18, 19, 20, 21]
"squared.normalized.residuals"        => [5.53742e-5, 2.42977e-5, 2.36066e-6, 2.77706e-6, 1.07985e-7…
"residuals"                           => [2.5348, 1.67908, 0.523367, 0.567651, 0.111936, -0.343779, …
"basic.subset"                        => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  15, 16, 17, 18, 19, 20, …

References

Billor, Nedret, Samprit Chatterjee, and Ali S. Hadi. "A re-weighted least squares method for robust regression estimation." American journal of mathematical and management sciences 26.3-4 (2006): 229-252.

LinRegOutliers.bchplot — Method

bchplot(setting::RegressionSetting; alpha=0.05, maxiter=1000, epsilon=0.00001)

Perform the Billor & Chatterjee & Hadi (2006) algorithm and generates outlier plot for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

alpha::Float64: Optional argument of the probability of rejecting the null hypothesis.

maxiter::Int: Maximum number of iterations for calculating iterative weighted least squares estimates.

epsilon::Float64: Accuracy for determining convergency.

References

Billor, Nedret, Samprit Chatterjee, and Ali S. Hadi. "A re-weighted least squares method for robust regression estimation." American journal of mathematical and management sciences 26.3-4 (2006): 229-252.

LinRegOutliers.ccf — Method

ccf(X, y; starting_lambdas = nothing)

Perform signed gradient descent for clipped convex functions for a given regression setting.

Arguments

X::Array{Float64, 2}: Design matrix of the linear model.

y::Array{Float64, 1}: Response vector of the linear model.

starting_lambdas::Array{Float64,1}: Starting values of weighting parameters used by signed gradient descent.

alpha::Float64: Loss at which a point is labeled as an outlier. If unspecified, will be chosen as p*mean(residuals.^2), where residuals are OLS residuals.

p::Float64: Points that have squared OLS residual greater than p times the mean squared OLS residual are considered outliers.

max_iter::Int64: Maximum number of iterations to run signed gradient descent.

beta::Float64: Step size parameter.

tol::Float64: Tolerance below which convergence is declared.

References

Barratt, S., Angeris, G. & Boyd, S. Minimizing a sum of clipped convex functions. Optim Lett 14, 2443–2459 (2020). https://doi.org/10.1007/s11590-020-01565-4

LinRegOutliers.ccf — Method

ccf(setting; starting_lambdas = nothing)

Perform signed gradient descent for clipped convex functions for a given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

starting_lambdas::Array{Float64,1}: Starting values of weighting parameters used by signed gradient descent.

alpha::Float64: Loss at which a point is labeled as an outlier (points with loss ≥ alpha will be called outliers).

max_iter::Int64: Maximum number of iterations to run signed gradient descent.

beta::Float64: Step size parameter.

tol::Float64: Tolerance below which convergence is declared.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> ccf(reg0001)
Dict{Any,Any} with 4 entries:
  "betas"     => [-63.4816, 1.30406]
  "outliers"  => [15, 16, 17, 18, 19, 20]
  "lambdas"   => [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0  …  2.77556e-17, 2.77556e-17, 0…
  "residuals" => [-2.67878, -1.67473, -0.37067, -0.266613, 0.337444, 0.941501, 1.44556, 2.04962, 1…

References

Barratt, S., Angeris, G. & Boyd, S. Minimizing a sum of clipped convex functions. Optim Lett 14, 2443–2459 (2020). https://doi.org/10.1007/s11590-020-01565-4

LinRegOutliers.cga — Method

Performs a CGA (Compact Genetic Algorithm) search for minimization of an objective function. In the example below, the objective function is to minimize sum of bits of a binary vector. The search method results the optimum vector of [0, 0, ..., 0] where the objective function is zero.

Examples

julia> function f(x)
           return sum(x)
       end
f (generic function with 1 method)
julia> cga(chsize = 10, costfunction = f, popsize = 100)
10-element Array{Bool,1}:
 0
 0
 0
 0
 0
 0
 0
 0
 0
 0

LinRegOutliers.cgasample — Method

Generates a binary array of values using a probability vector. Each single element of the probability vector is the probability of bit having the value of 1. When the probability vector is [1, 1, 1, ..., 1] then the sampled vector is [1.0, 1.0, 1.0, ..., 1.0] whereas it is [0.0, 0.0, 0.0, ..., 0.0] when the probability vector is a vector of zeros. The CGA (compact genetic algorithms) search is started using the probability vector of [0.5, 0.5, 0.5, ..., 0.5] which produces random vectors of either zeros or ones.

Examples

julia> sample([1, 1, 1, 1, 1])
5-element Array{Bool,1}:
 1
 1
 1
 1
 1
julia> cgasample(ones(10) * 0.5)
10-element Array{Bool,1}:
 1
 1
 1
 1
 0
 0
 0
 1
 1
 0

LinRegOutliers.cm97 — Method

cm97(setting; maxiter = 1000)

Perform the Chatterjee and Mächler (1997) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> myreg = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> result = cm97(myreg)
Dict{String,Any} with 3 entries:
  "betas"      => [-37.0007, 0.839285, 0.632333, -0.113208]
  "iterations" => 22
  "converged"  => true

References

Chatterjee, Samprit, and Martin Mächler. "Robust regression: A weighted least squares approach." Communications in Statistics-Theory and Methods 26.6 (1997): 1381-1394.

LinRegOutliers.compute_t_distance — Method

    compute_t_distance(X, y, subset)

This function computes the t distance for each point and returns the distance vector.

Arguments:

X: The multivariate data matrix.

y: The output vector

subset: The vector which denotes the points inside the subset, used to scale the residuals accordingly.

LinRegOutliers.cooks — Method

cooks(setting)

Calculate Cook distances for all observations in a regression setting.

Arguments

setting::RegressionSetting: A regression setting object.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);

julia> cooks(reg)
24-element Array{Float64,1}:
 0.005344774190779822
 0.0017088194691033689
 0.00016624914057962608
 3.1644452583114795e-5
 0.0005395058666404081
 0.0014375008774859539
 0.0024828140956511258
 0.0036279720445167277
 0.004357605989540906
 0.005288503758364767
 0.006313578057565415
 0.0076561205696857254
 0.009568574875389256
 0.009970039008782357
 0.02610396373381051
 0.029272523880917646
 0.05091236198400663
 0.08176555044049343
 0.14380266904640235
 0.26721539425047447
 0.051205153558783356
 0.13401084683481085
 0.16860324592350226
 0.2172819114905912

References

Cook, R. Dennis. "Detection of influential observation in linear regression." Technometrics 19.1 (1977): 15-18.

LinRegOutliers.coordinatwisemedians — Method

coordinatwisemedians(datamat)

Return vector of medians of each column in a matrix.

Arguments

datamat::Array{Float64, 2}: A matrix.

Example

julia> mat = [1.0 2.0; 3.0 4.0; 5.0 6.0]
3×2 Array{Float64,2}:
 1.0  2.0
 3.0  4.0
 5.0  6.0

julia> coordinatwisemedians(mat)
2-element Array{Float64,1}:
 3.0
 4.0

LinRegOutliers.covratio — Method

covratio(setting, omittedIndex)

Apply covariance ratio diagnostic for a given regression setting and observation index.

Arguments

setting::RegressionSetting: A regression setting object.

omittedIndex::Int: Index of the omitted observation.

Example

julia> setting = createRegressionSetting(@formula(calls ~ year), phones);
julia> covratio(setting, 1)
1.2945913799871505

LinRegOutliers.createRegressionSetting — Method

createRegressionSetting(formula, data)

Create a regression setting for a given formula and data

Arguments

formula::FormulaTerm: A formula object describes the linear regression model.

data::DataFrame: DataFrame object holds the data

Notes

Implemented methods in this packages accepts linear models as RegressionSetting objects.
This objects holds the model formula and the data used in regression estimations.
createRegressionSetting is a helper function for creating RegressionSetting objects.

Examples

julia> setting = createRegressionSetting(@formula(calls ~ year), phones)
RegressionSetting(calls ~ year, 24×2 DataFrame
│ Row │ year  │ calls   │
│     │ Int64 │ Float64 │
├─────┼───────┼─────────┤
│ 1   │ 50    │ 4.4     │
│ 2   │ 51    │ 4.7     │
│ 3   │ 52    │ 4.7     │
│ 4   │ 53    │ 5.9     │
│ 5   │ 54    │ 6.6     │
│ 6   │ 55    │ 7.3     │
│ 7   │ 56    │ 8.1     │
│ 8   │ 57    │ 8.8     │
│ 9   │ 58    │ 10.6    │
│ 10  │ 59    │ 12.0    │
⋮
│ 14  │ 63    │ 21.2    │
│ 15  │ 64    │ 119.0   │
│ 16  │ 65    │ 124.0   │
│ 17  │ 66    │ 142.0   │
│ 18  │ 67    │ 159.0   │
│ 19  │ 68    │ 182.0   │
│ 20  │ 69    │ 212.0   │
│ 21  │ 70    │ 43.0    │
│ 22  │ 71    │ 24.0    │
│ 23  │ 72    │ 27.0    │
│ 24  │ 73    │ 29.0    │)

LinRegOutliers.dataimage — Method

dataimage(dataMatrix; distance = :mahalanobis)

Generate the Marchette & Solka (2003) data image for a given data matrix.

Arguments

dataMatrix::Array{Float64, 1}: Data matrix with dimensions n x p, where n is the number of observations and p is the number of variables.

distance::Symbol: Optional argument for the distance function.

Notes

distance is :mahalanobis by default, for the Mahalanobis distances. 
use 

    dataimage(mat, distance = :euclidean)

to use Euclidean distances.

Examples

julia> x1 = hbk[:,"x1"];
julia> x2 = hbk[:,"x2"];
julia> x3 = hbk[:,"x3"];
julia> mat = hcat(x1, x2, x3);
julia> dataimage(mat)

References

Marchette, David J., and Jeffrey L. Solka. "Using data images for outlier detection." Computational Statistics & Data Analysis 43.4 (2003): 541-552.

LinRegOutliers.designMatrix — Method

designMatrix(setting)

Return matrix of independent variables including the variable (ones) of the constanst term for a given regression setting.

Arguments

setting::RegressionSetting: A regression setting object.

Notes

Design matrix is a matrix which holds values of independent variables on its columns for a given linear regression model.

Examples

julia> setting = createRegressionSetting(@formula(calls ~ year), phones);
julia> designMatrix(setting)
24×2 Array{Float64,2}:
 1.0  50.0
 1.0  51.0
 1.0  52.0
 1.0  53.0
 1.0  54.0
 1.0  55.0
 1.0  56.0
 1.0  57.0
 1.0  58.0
 1.0  59.0
 1.0  60.0
 1.0  61.0
 1.0  62.0
 1.0  63.0
 1.0  64.0
 1.0  65.0
 1.0  66.0
 1.0  67.0
 1.0  68.0
 1.0  69.0
 1.0  70.0
 1.0  71.0
 1.0  72.0
 1.0  73.0

LinRegOutliers.dfbeta — Method

dfbeta(setting, omittedIndex)

Apply DFBETA diagnostic for a given regression setting and observation index.

Arguments

setting::RegressionSetting: A regression setting object.

omittedIndex::Int: Index of the omitted observation.

Example

julia> setting = createRegressionSetting(@formula(calls ~ year), phones);
julia> dfbeta(setting, 1)
2-element Array{Float64,1}:
  9.643915678524024
 -0.14686166007904422

LinRegOutliers.dffit — Method

dffit(setting, i)

Calculate the effect of the ith observation on the linear regression fit.

Arguments

setting::RegressionSetting: A regression setting object.

i::Int: Index of the observation.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);
julia> dffit(reg, 1)
2.3008326745719785

julia> dffit(reg, 15)
2.7880619386124295

julia> dffit(reg, 16)
3.1116532421969794

julia> dffit(reg, 17)
4.367981450347031

julia> dffit(reg, 21)
-5.81610150322166

References

Belsley, David A., Edwin Kuh, and Roy E. Welsch. Regression diagnostics: Identifying influential data and sources of collinearity. Vol. 571. John Wiley & Sons, 2005.

LinRegOutliers.dffit — Method

dffit(setting)

Calculate dffit for all observations.

Arguments

setting::RegressionSetting: A regression setting object.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);

julia> dffit(reg)
24-element Array{Float64,1}:
   2.3008326745719785
   1.2189579001467337
   0.35535667547543426
  -0.14458523141740898
  -0.5558346324846752
  -0.8441316814464983
  -1.0329184407957257
  -1.16600692151232
  -1.2005633711667656
  -1.2549187193476428
  -1.3195581500053777
  -1.42383876236147
  -1.5917690629803474
  -1.6582086833534504
   2.7880619386124295
   3.1116532421969794
   4.367981450347031
   5.927603041427858
   8.442860517217582
  12.370243663029527
  -5.81610150322166
 -10.089153963127842
 -12.10803256546825
 -14.67006851119936

References

Belsley, David A., Edwin Kuh, and Roy E. Welsch. Regression diagnostics: Identifying influential data and sources of collinearity. Vol. 571. John Wiley & Sons, 2005.

LinRegOutliers.dominates — Method

dominates(p1::Array, p2::Array)

Return true if each element in p1 is not less than the corresponding element in p2 and at least one element in p1 is bigger than the corresponding element in p2.

Arguments

p1::Array: Numeric array of n elements.

p2::Array: Numeric array of n elements.

Examples

julia> dominates([1,2,3], [1,2,1])
true

julia> dominates([0,0,0,0], [1,0,0,0])
false

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

Deb, Kalyanmoy, et al. "A fast elitist non-dominated sorting genetic algorithm for multi-objective optimization: NSGA-II." International conference on parallel problem solving from nature. Springer, Berlin, Heidelberg, 2000.

LinRegOutliers.euclideanDistances — Method

euclideanDistances(dataMatrix)

Calculate Euclidean distances between pairs.

Arguments

dataMatrix::Array{Float64, 1}: Data matrix with dimensions n x p, where n is the number of observations and p is the number of variables.

Notes

This is the helper function for the dataimage() function defined in Marchette & Solka (2003).

References

Marchette, David J., and Jeffrey L. Solka. "Using data images for outlier detection." Computational Statistics & Data Analysis 43.4 (2003): 541-552.

LinRegOutliers.find_minimum_nonzero — Method

find_minimum_nonzero(arr)

Return minimum of numbers greater than zero.

Arguments

arr::Array{Float64, 1}: A function that takes a one dimensional array as argument.

Example

julia> find_minimum_nonzero([0.0, 0.0, 5.0, 1.0])
1.0

LinRegOutliers.galts — Method

galts(setting)

Perform Satman(2012) algorithm for estimating LTS coefficients.

Arguments

setting: A regression setting object.

Examples

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);
julia> galts(reg)
Dict{Any,Any} with 3 entries:
  "betas"       => [-56.5219, 1.16488]
  "best.subset" => [3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 23, 24]
  "objective"   => 3.43133

References

Satman, M. Hakan. "A genetic algorithm based modification on the lts algorithm for large data sets." Communications in Statistics-Simulation and Computation 41.5 (2012): 644-652.

LinRegOutliers.hadi1992 — Method

hadi1992(multivariateData)

Perform Hadi (1992) algorithm for a given multivariate data.

# Arguments -multivariate::Array{Float64, 2}: Multivariate data.

Examples

julia> multidata = hcat(hbk.x1, hbk.x2, hbk.x3);

julia> hadi1992(multidata)
Dict{Any,Any} with 3 entries:
  "outliers"              => [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
  "criticial.chi.squared" => 7.81473
  "rth.robust.distance"   => 5.04541

# Reference Hadi, Ali S. "Identifying multiple outliers in multivariate data." Journal of the Royal Statistical Society: Series B (Methodological) 54.3 (1992): 761-771.

LinRegOutliers.hadi1992_handle_singularity — Method

hadi1992_handle_singularity(S)

Perform the sub-algorithm of handling singularity defined in Hadi (1992).

# Arguments

S::Array{Float64, 2}: A covariance matrix.

# Reference Hadi, Ali S. "Identifying multiple outliers in multivariate data." Journal of the Royal Statistical Society: Series B (Methodological) 54.3 (1992): 761-771.

LinRegOutliers.hadi1994 — Method

hadi1994(multivariateData)

Perform Hadi (1994) algorithm for a given multivariate data.

# Arguments -multivariate::Array{Float64, 2}: Multivariate data.

Examples

julia> multidata = hcat(hbk.x1, hbk.x2, hbk.x3);

julia> hadi1994(multidata)
Dict{Any,Any} with 3 entries:
  "outliers"              => [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
  "critical.chi.squared" => 7.81473
  "rth.robust.distance"   => 5.04541

# Reference Hadi, Ali S. "A modification of a method for the dedection of outliers in multivariate samples" Journal of the Royal Statistical Society: Series B (Methodological) 56.2 (1994): 393-396.

LinRegOutliers.hadimeasure — Method

hadimeasure(setting; c = 2.0)

Apply Hadi's regression diagnostic for a given regression setting

Arguments

setting::RegressionSetting: A regression setting object.

c::Float64: Critical value selected between 2.0 - 3.0. The default is 2.0.

Example

julia> setting = createRegressionSetting(@formula(calls ~ year), phones);
julia> hadimeasure(setting)

References

Chatterjee, Samprit and Hadi, Ali. Regression Analysis by Example. 5th ed. N.p.: John Wiley & Sons, 2012.

LinRegOutliers.hatmatrix — Method

hatmatrix(setting)

Calculate Hat matrix of dimensions n x n for a given regression setting with n observations.

Arguments

setting::RegressionSetting: A regression setting object.

Examples

```julia-repl julia> reg = createRegressionSetting(@formula(calls ~ year), phones); julia> size(hatmatrix(reg))

(24, 24)

LinRegOutliers.hs93 — Method

hs93(setting; alpha = 0.05, basicsubsetindices = nothing)

Perform the Hadi & Simonoff (1993) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

alpha::Float64: Optional argument of the probability of rejecting the null hypothesis.

basicsubsetindices::Array{Int, 1}: Initial basic subset, by default, the algorithm creates an initial set of clean observations.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> hs93(reg0001)
Dict{Any,Any} with 3 entries:
  "outliers" => [14, 15, 16, 17, 18, 19, 20, 21]
  "t"        => -3.59263
  "d"        => [2.04474, 1.14495, -0.0633255, 0.0632934, -0.354349, -0.766818, -1.06862, -1.47638, -0.7…

References

Hadi, Ali S., and Jeffrey S. Simonoff. "Procedures for the identification of multiple outliers in linear models." Journal of the American Statistical Association 88.424 (1993): 1264-1272.

LinRegOutliers.hs93basicsubset — Method

hs93basicsubset(setting, initialindices)

Perform the Hadi & Simonoff (1993) algorithm's second part for a given regression setting. The returned array of indices are indices of clean subset of length h where h is at least the half of the number of observations. h is set to integer part of (n + p - 1) / 2.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

initialindices::Array{Int, 1}: (p + 1) subset of clean observations.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> initialsetindices = hs93initialset(reg0001)
3-element Array{Int64,1}:
 4
 3
 5
 julia> hs93basicsubset(reg0001, initialsetindices)
12-element Array{Int64,1}:
  5
  9
 10
  3
  6
  4
  7
 22
 11
  8
 12
 13

References

Hadi, Ali S., and Jeffrey S. Simonoff. "Procedures for the identification of multiple outliers in linear models." Journal of the American Statistical Association 88.424 (1993): 1264-1272.

LinRegOutliers.hs93initialset — Method

hs93initialset(setting)

Perform the Hadi & Simonoff (1993) algorithm's first part for a given regression setting. The returned array of indices are indices of clean subset length of p + 1 where p is the number of regression parameters.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> hs93initialset(reg0001)
3-element Array{Int64,1}:
 4
 3
 5

References

Hadi, Ali S., and Jeffrey S. Simonoff. "Procedures for the identification of multiple outliers in linear models." Journal of the American Statistical Association 88.424 (1993): 1264-1272.

LinRegOutliers.imon2005 — Method

imon2005(setting)

Perform the Imon 2005 algorithm for a given regression setting.

# Arguments

setting::RegressionSetting: A regression setting.

Notes

The implementation uses LTS rather than LMS as suggested in the paper.

# Reference A. H. M. Rahmatullah Imon (2005) Identifying multiple influential observations in linear regression, Journal of Applied Statistics, 32:9, 929-946, DOI: 10.1080/02664760500163599

LinRegOutliers.initial_basic_subset_multivariate_data — Method

    initial_basic_subset_multivariate_data(X, m, method="mahalanobis")

This function returns the m-subset according to algorithm #1 for multivariate data. Two methods V1 and V2 are defined in the paper which use Mahalanobis distance or distance from the coordinate-wise median. The m subset returned is guaranteed to be full rank.

Arguments

X: The multivariate matrix where each row is a data point

m: The number of points to include in the initial subset

method: The distance method to use for selecting the points for initial subset

LinRegOutliers.iterateCSteps — Method

iterateCSteps(setting, subsetindices, h)

Perform a concentration step for a given subset of a regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

subsetindices::Array{Int, 1}: Indicies of observations in the initial subset.

h::Int: A constant at least half of the number of observations.

Notes

This function is a helper for the lts function. A concentration step starts with a 
initial subset. The size of the subset is enlarged to h, a constant at least half of the 
observations. Please refer to the citation given below.

References

Rousseeuw, Peter J., and Katrien Van Driessen. "An algorithm for positive-breakdown regression based on concentration steps." Data Analysis. Springer, Berlin, Heidelberg, 2000. 335-346.

LinRegOutliers.jacknifedS — Method

jacknifedS(setting, omittedIndices)

Calculate Jacknife standard error in which the given indices are omitted from the data.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

omittedIndices::Array{Int, 1}: Indices of omitted variables.

References

Peña, Daniel, and Victor J. Yohai. "The detection of influential subsets in linear regression by using an influence matrix." Journal of the Royal Statistical Society: Series B (Methodological) 57.1 (1995): 145-156.

LinRegOutliers.jacknifedS — Method

jacknifedS(setting, k)

Estimate standard error of regression with the kth observation is dropped.

# Arguments

setting::RegressionSetting: A regression setting object.

k::Int: Index of the omitted observation.

# Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);
julia> jacknifedS(reg, 2)
57.518441664761035

julia> jacknifedS(reg, 15)
56.14810222161477

LinRegOutliers.ks89 — Method

ks89(setting; alpha = 0.05)

Perform the Kianifard & Swallow (1989) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

alpha::Float64: Optional argument of the probability of rejecting the null hypothesis.

Examples

julia> reg0001 = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> ks89(reg0001)
2-element Array{Int64,1}:
  4
 21

References

Kianifard, Farid, and William H. Swallow. "Using recursive residuals, calculated on adaptively-ordered observations, to identify outliers in linear regression." Biometrics (1989): 571-585.

LinRegOutliers.ks89RecursiveResidual — Method

ks89RecursiveResidual(setting; indices, k)

Calculate recursive residual for the given regression setting and observation.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

indices::ArrayInt,1: Indices of observations used in the linear model.

k::Int: Observation indice the recursive residual is calculated for.

Notes

This is a helper function for the ks89 function and it is not directly used.

References

Kianifard, Farid, and William H. Swallow. "Using recursive residuals, calculated on adaptively-ordered observations, to identify outliers in linear regression." Biometrics (1989): 571-585.

LinRegOutliers.lad — Method

lad(X, y; starting_betas = nothing)

Perform Least Absolute Deviations regression for a given regression setting.

Arguments

X::Array{Float64, 2}: Design matrix of the linear model.

y::Array{Float64, 1}: Response vector of the linear model.

starting_betas::Array{Float64,1}: Starting values of parameter estimations that fed to local search optimizer.

LinRegOutliers.lad — Method

lad(setting; starting_betas = nothing)

Perform Least Absolute Deviations regression for a given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

starting_betas::Array{Float64,1}: Starting values of parameter estimations that fed to local search optimizer.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> lad(reg0001)
Dict{Any,Any} with 2 entries:
  "betas"     => [-57.3269, 1.19155]
  "residuals" => [2.14958, 1.25803, 0.0664872, 0.0749413, -0.416605, -0.90815, -1.2997, -1.79124,…

LinRegOutliers.lms — Method

lms(setting; iters = nothing, crit = 2.5)

Perform Least Median of Squares regression estimator with random sampling.

Arguments

setting::RegressionSetting: A regression setting object.

iters::Int: Number of random samples.

crit::Float64: Critical value for standardized residuals.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);

julia> lms(reg)
Dict{Any,Any} with 6 entries:
  "stdres"    => [2.28328, 1.55551, 0.573308, 0.608843, 0.220321, -0.168202, -0.471913, -0.860435, -0.31603, -0.110871  …  85.7265, 88.9849, 103.269, 116.705, 135.229, 159.69,…
  "S"         => 1.17908
  "outliers"  => [14, 15, 16, 17, 18, 19, 20, 21]
  "objective" => 0.515348
  "coef"      => [-56.1972, 1.1581]
  "crit"      => 2.5

References

Rousseeuw, Peter J. "Least median of squares regression." Journal of the American statistical association 79.388 (1984): 871-880.

LinRegOutliers.lta — Method

lta(X, y; exact = false)

Perform the Hawkins & Olive (1999) algorithm (Least Trimmed Absolute Deviations) for the given regression setting.

Arguments

X::Array{Float64, 2}: Design matrix of linear regression model.

y::Array{Float64, 1}: Response vector of linear regression model.

exact::Bool: Consider all possible subsets of p or not where p is the number of regression parameters.

References

Hawkins, Douglas M., and David Olive. "Applications and algorithms for least trimmed sum of absolute deviations regression." Computational Statistics & Data Analysis 32.2 (1999): 119-134.

LinRegOutliers.lta — Method

lta(setting; exact = false)

Perform the Hawkins & Olive (1999) algorithm (Least Trimmed Absolute Deviations) for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

exact::Bool: Consider all possible subsets of p or not where p is the number of regression parameters.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> lta(reg0001)
Dict{Any,Any} with 2 entries:
  "betas"     => [-55.5, 1.15]
  "objective" => 5.7

julia> lta(reg0001, exact = true)
Dict{Any,Any} with 2 entries:
  "betas"     => [-55.5, 1.15]
  "objective" => 5.7  

References

Hawkins, Douglas M., and David Olive. "Applications and algorithms for least trimmed sum of absolute deviations regression." Computational Statistics & Data Analysis 32.2 (1999): 119-134.

LinRegOutliers.lts — Method

lts(setting; iters, crit)

Perform the Fast-LTS (Least Trimmed Squares) algorithm for a given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

iters::Int: Number of iterations.

crit::Float64: Critical value.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);
julia> lts(reg)
Dict{Any,Any} with 6 entries:
  "betas"            => [-56.5219, 1.16488]
  "S"                => 1.10918
  "hsubset"          => [11, 10, 5, 6, 23, 12, 13, 9, 24, 7, 3, 4, 8]
  "outliers"         => [14, 15, 16, 17, 18, 19, 20, 21]
  "scaled.residuals" => [2.41447, 1.63472, 0.584504, 0.61617, 0.197052, -0.222066, -0.551027, -0.970146, -0.397538, -0.185558  …  …
  "objective"        => 3.43133

References

Rousseeuw, Peter J., and Katrien Van Driessen. "An algorithm for positive-breakdown regression based on concentration steps." Data Analysis. Springer, Berlin, Heidelberg, 2000. 335-346.

LinRegOutliers.mahalanobisBetweenPairs — Method

mahalanobisBetweenPairs(dataMatrix)

Calculate Mahalanobis distances between pairs.

Arguments

dataMatrix::Array{Float64, 1}: Data matrix with dimensions n x p, where n is the number of observations and p is the number of variables.

Notes

Differently from Mahalabonis distances, this function calculates Mahalanobis distances between 
pairs, rather than the distances to center of the data. This is the helper function for the 
dataimage() function defined in Marchette & Solka (2003).

References

Marchette, David J., and Jeffrey L. Solka. "Using data images for outlier detection." Computational Statistics & Data Analysis 43.4 (2003): 541-552.

LinRegOutliers.mahalanobisSquaredMatrix — Method

mahalanobisSquaredMatrix(data::DataFrame; meanvector=nothing, covmatrix=nothing)

Calculate Mahalanobis distances.

Arguments

data::DataFrame: A DataFrame object of the multivariate data.

meanvector::Array{Float64, 1}: Optional mean vector of variables.

covmatrix::Array{Float64, 2}: Optional covariance matrix of data.

# References Mahalanobis, Prasanta Chandra. "On the generalized distance in statistics." National Institute of Science of India, 1936.

LinRegOutliers.mcd — Method

mcd(data; alpha = 0.01)

Performs the Minimum Covariance Determinant algorithm for a robust covariance matrix.

Arguments

data::DataFrame: Multivariate data.

alpha::Float64: Probability for quantiles of Chi-Squared statistic.

# Notes Algorithm is implemented using concentration steps as described in the reference paper. However, details about number of iterations are slightly different.

References

Rousseeuw, Peter J., and Katrien Van Driessen. "A fast algorithm for the minimum covariance determinant estimator." Technometrics 41.3 (1999): 212-223.

LinRegOutliers.midlist — Method

midlist(n::Int, p::Int)

Return p indices in the middle of 1:n.

Arguments

n::Int: Number of observations.

p::Int: Number of elements in the middle of indices.

Notes

If n is even and p is odd then p + 1 observation indices are returned.

Examples

julia> midlist(10,2)
2-element Array{Int64,1}:
 5
 6

julia> midlist(10,3)
4-element Array{Int64,1}:
 4
 5
 6
 7

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

LinRegOutliers.mve — Method

mve(data; alpha = 0.01)

Performs the Minimum Volume Ellipsoid algorithm for a robust covariance matrix.

Arguments

data::DataFrame: Multivariate data.

alpha::Float64: Probability for quantiles of Chi-Squared statistic.

References

Van Aelst, Stefan, and Peter Rousseeuw. "Minimum volume ellipsoid." Wiley Interdisciplinary Reviews: Computational Statistics 1.1 (2009): 71-82.

LinRegOutliers.mveltsplot — Method

mveltsplot(setting; alpha = 0.05, showplot = true)

Generate MVE - LTS plot for visual detecting of regression outliers.

Arguments

setting::RegressionSetting: A regression setting object.

alpha::Float64: Probability for quantiles of Chi-Squared statistic.

showplot::Bool: Whether a plot is shown or only return statistics.

References

Van Aelst, Stefan, and Peter Rousseeuw. "Minimum volume ellipsoid." Wiley Interdisciplinary Reviews: Computational Statistics 1.1 (2009): 71-82.

LinRegOutliers.ndsranks — Method

ndsranks(data)

Sort multidimensional data usin non-dominated sorting algorithm.

Arguments

data::Matrix: n x k matrix of observations where n is number of observations and k is number of variables.

Examples

julia> datamat = convert(Matrix, hbk)
75×4 Array{Float64,2}:
 10.1  19.6  28.3   9.7
  9.5  20.5  28.9  10.1
 10.7  20.2  31.0  10.3
  9.9  21.5  31.7   9.5
 10.3  21.1  31.1  10.0
 10.8  20.4  29.2  10.0
 10.5  20.9  29.1  10.8
  9.9  19.6  28.8  10.3
  9.7  20.7  31.0   9.6
  9.3  19.7  30.3   9.9
 11.0  24.0  35.0  -0.2
 12.0  23.0  37.0  -0.4
  ⋮                
  2.8   3.0   2.9  -0.5
  2.0   0.7   2.7   0.6
  0.2   1.8   0.8  -0.9
  1.6   2.0   1.2  -0.7
  0.1   0.0   1.1   0.6
  2.0   0.6   0.3   0.2
  1.0   2.2   2.9   0.7
  2.2   2.5   2.3   0.2
  0.6   2.0   1.5  -0.2
  0.3   1.7   2.2   0.4
  0.0   2.2   1.6  -0.9
  0.3   0.4   2.6   0.2

julia> ndsranks(datamat)
75-element Array{Int64,1}:
 61
 61
 64
 61
 64
 62
 64
 61
 61
 61
 30
 23
  ⋮
 11
  7
  0
  2
  0
  0
 12
 14
  4
  1
  0
  0

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

Deb, Kalyanmoy, et al. "A fast elitist non-dominated sorting genetic algorithm for multi-objective optimization: NSGA-II." International conference on parallel problem solving from nature. Springer, Berlin, Heidelberg, 2000.

LinRegOutliers.ndsranks — Method

ndsranks(data)

Sort multidimensional data usin non-dominated sorting algorithm.

Arguments

data::DataFrame: DataFrame of variables.

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

Deb, Kalyanmoy, et al. "A fast elitist non-dominated sorting genetic algorithm for multi-objective optimization: NSGA-II." International conference on parallel problem solving from nature. Springer, Berlin, Heidelberg, 2000.

LinRegOutliers.py95 — Method

py95(setting)

Perform the Pena & Yohai (1995) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> reg0001 = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk);
julia> py95(reg0001)
ict{Any,Any} with 2 entries:
  "outliers"       => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
  "suspected.sets" => Set([[14, 13], [43, 54, 24, 38, 22], [6, 10], [14, 7, 8, 3, 10, 2, 5, 6, 1, 9, 4…

References

Peña, Daniel, and Victor J. Yohai. "The detection of influential subsets in linear regression by using an influence matrix." Journal of the Royal Statistical Society: Series B (Methodological) 57.1 (1995): 145-156.

LinRegOutliers.py95ProcessEigenVector — Method

py95ProcessEigenVector(v)

Process eigen vectors of EDHDE matrix as defined in the Pena & Yohai (1995) algorithm.

Arguments

v::Array{Float64, 1}: Eigen vector of EDHDE matrix.

References

Peña, Daniel, and Victor J. Yohai. "The detection of influential subsets in linear regression by using an influence matrix." Journal of the Royal Statistical Society: Series B (Methodological) 57.1 (1995): 145-156.

LinRegOutliers.py95SuspectedObservations — Method

py95SuspectedObservations(setting)

Determine suspected observations (outliers) as defined in the Pena & Yohai (1995) algorithm.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

References

Peña, Daniel, and Victor J. Yohai. "The detection of influential subsets in linear regression by using an influence matrix." Journal of the Royal Statistical Society: Series B (Methodological) 57.1 (1995): 145-156.

LinRegOutliers.ransac — Method

ransac(setting; t, w=0.5, m=0, k=0, d=0, confidence=0.99)

Run the RANSAC (1981) algorithm for the given regression setting

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and a dataset.

t::Float64: The threshold distance of a sample point to the regression hyperplane to determine if it fits the model well.

w::Float64: The probability of a sample point being inlier, default=0.5.

m::Int: The number of points to sample to estimate the model parameter for each iteration. If set to 0, defaults to picking p points which is the minimum required.

k::Int: The number of iterations to run. If set to 0, is calculated according to the formula given in the paper based on outlier probability and the sample set size.

d::Int: The number of close data points required to accept the model. Defaults to number of data points multiplied by inlier ratio.

confidence::Float64: Required to determine the number of optimum iterations if k is not specified.

Examples

julia> df = DataFrame(y=[0,1,2,3,3,4,10], x=[0,1,2,2,3,4,2])
julia> reg = createRegressionSetting(@formula(y ~ x), df)
julia> ransac(reg, t=0.8, w=0.85)
1-element Array{Int64,1}:
 7

References

Martin A. Fischler & Robert C. Bolles (June 1981). "Random Sample Consensus: A Paradigm for Model Fitting with Applications to Image Analysis and Automated Cartography" Comm. ACM. 24 (6): 381–395.

LinRegOutliers.responseVector — Method

responseVector(setting)

Return vector of dependent variable of a given regression setting.

Arguments

setting::RegressionSetting: A regression setting object.

Examples

julia> setting = createRegressionSetting(@formula(calls ~ year), phones);
julia> responseVector(setting)
24-element Array{Float64,1}:
   4.4
   4.7
   4.7
   5.9
   6.6
   7.3
   8.1
   8.8
  10.6
  12.0
  13.5
  14.9
  16.1
  21.2
 119.0
 124.0
 142.0
 159.0
 182.0
 212.0
  43.0
  24.0
  27.0
  29.0

LinRegOutliers.satman2013 — Method

satman2013(setting)

Perform Satman (2013) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> eg0001 = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk);
julia> satman2013(reg0001)
Dict{Any,Any} with 1 entry:
  "outliers" => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 47]

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

LinRegOutliers.satman2015 — Method

satman2013(setting)

Perform Satman (2015) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> eg0001 = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk);
julia> satman2013(reg0001)
Dict{Any,Any} with 1 entry:
  "outliers" => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 14, 47]

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

LinRegOutliers.select_subset — Method

    select_subset(X, m, distances)

This function returns the list of indices which have the least distances as given in the distances array. It also guarantees that at least m indices are returned and that the selected indices have the full rank.

Arguments

X: The multivariate matrix where each row is a data point.

m: The minimum number of points to include in the subset indices.

distances: The distances vector used for selecting minumum distance indices.

LinRegOutliers.smr98 — Method

smr98(setting)

Perform the Sebert, Monthomery and Rollier (1998) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> smr98(reg0001)
10-element Array{Int64,1}:
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24

References

Sebert, David M., Douglas C. Montgomery, and Dwayne A. Rollier. "A clustering algorithm for identifying multiple outliers in linear regression." Computational statistics & data analysis 27.4 (1998): 461-484.

LinRegOutliers.studentizedResiduals — Method

studentizedResiduals(setting)

Calculate Studentized residuals for a given regression setting.

# Arguments:

setting::RegressionSetting: A regression setting object.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);

julia> studentizedResiduals(reg)
24-element Array{Float64,1}:
  0.2398783264505892
  0.1463945666608097
  0.04934549995087145
 -0.023289236798461784
 -0.10408303320973748
 -0.18382934382804111
 -0.2609395640240455
 -0.33934473417314376
 -0.3973205657179429
 -0.46258080183149236
 -0.5261488085924144
 -0.5918396227060093
 -0.6616423337899147
 -0.6611792918262785
  1.0277190922689816
  1.0297863954540103
  1.2712201589839855
  1.4974523565936426
  1.8386296155264197
  2.316394853333409
 -0.9368354141338643
 -1.4009989983319822
 -1.4541520919831887
 -1.529459974327181